// training_frontend/src/api/models/Exercise.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Training service
 * Backend for training app
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { MuscleGroup } from './MuscleGroup';
import {
    MuscleGroupFromJSON,
    MuscleGroupFromJSONTyped,
    MuscleGroupToJSON,
    MuscleGroupToJSONTyped,
} from './MuscleGroup';

/**
 * 
 * @export
 * @interface Exercise
 */
export interface Exercise {
    /**
     * 
     * @type {string}
     * @memberof Exercise
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof Exercise
     */
    name: string;
    /**
     * 
     * @type {boolean}
     * @memberof Exercise
     */
    bodyWeight: boolean;
    /**
     * 
     * @type {Array<MuscleGroup>}
     * @memberof Exercise
     */
    muscleGroups: Array<MuscleGroup>;
}

/**
 * Check if a given object implements the Exercise interface.
 */
export function instanceOfExercise(value: object): value is Exercise {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('bodyWeight' in value) || value['bodyWeight'] === undefined) return false;
    if (!('muscleGroups' in value) || value['muscleGroups'] === undefined) return false;
    return true;
}

export function ExerciseFromJSON(json: any): Exercise {
    return ExerciseFromJSONTyped(json, false);
}

export function ExerciseFromJSONTyped(json: any, ignoreDiscriminator: boolean): Exercise {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
        'bodyWeight': json['bodyWeight'],
        'muscleGroups': ((json['muscleGroups'] as Array<any>).map(MuscleGroupFromJSON)),
    };
}

export function ExerciseToJSON(json: any): Exercise {
    return ExerciseToJSONTyped(json, false);
}

export function ExerciseToJSONTyped(value?: Exercise | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'bodyWeight': value['bodyWeight'],
        'muscleGroups': ((value['muscleGroups'] as Array<any>).map(MuscleGroupToJSON)),
    };
}