// training_frontend/src/api/models/WorkoutExercise.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Training service
 * Backend for training app
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Exercise } from './Exercise';
import {
    ExerciseFromJSON,
    ExerciseFromJSONTyped,
    ExerciseToJSON,
    ExerciseToJSONTyped,
} from './Exercise';

/**
 * 
 * @export
 * @interface WorkoutExercise
 */
export interface WorkoutExercise {
    /**
     * 
     * @type {string}
     * @memberof WorkoutExercise
     */
    id: string;
    /**
     * 
     * @type {Exercise}
     * @memberof WorkoutExercise
     */
    exercise: Exercise;
    /**
     * 
     * @type {number}
     * @memberof WorkoutExercise
     */
    sets: number;
    /**
     * 
     * @type {Array<number>}
     * @memberof WorkoutExercise
     */
    repRange: Array<number>;
}

/**
 * Check if a given object implements the WorkoutExercise interface.
 */
export function instanceOfWorkoutExercise(value: object): value is WorkoutExercise {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('exercise' in value) || value['exercise'] === undefined) return false;
    if (!('sets' in value) || value['sets'] === undefined) return false;
    if (!('repRange' in value) || value['repRange'] === undefined) return false;
    return true;
}

export function WorkoutExerciseFromJSON(json: any): WorkoutExercise {
    return WorkoutExerciseFromJSONTyped(json, false);
}

export function WorkoutExerciseFromJSONTyped(json: any, ignoreDiscriminator: boolean): WorkoutExercise {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'exercise': ExerciseFromJSON(json['exercise']),
        'sets': json['sets'],
        'repRange': json['repRange'],
    };
}

export function WorkoutExerciseToJSON(json: any): WorkoutExercise {
    return WorkoutExerciseToJSONTyped(json, false);
}

export function WorkoutExerciseToJSONTyped(value?: WorkoutExercise | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'exercise': ExerciseToJSON(value['exercise']),
        'sets': value['sets'],
        'repRange': value['repRange'],
    };
}